+++ src/main/java/parkinglot/Car.java
@@ -1,0 +1,4 @@
+package parkinglot;
+
+public class Car {
+}
