--- src/main/java/parkinglot/beta/Car.java
@@ -1,4 +1,0 @@
-package parkinglot.beta;
-
-public class Car {
-}
