+++ src/main/java/parkinglot/exceptions/InvalidReceiptException.java
@@ -1,4 +1,4 @@
-package parkinglot.beta.exceptions;
+package parkinglot.exceptions;
 
 public class InvalidReceiptException extends RuntimeException {
     public InvalidReceiptException(String message) {
