+++ src/main/java/parkinglot/exceptions/NoSpaceInParkingLotException.java
@@ -1,4 +1,4 @@
-package parkinglot.beta.exceptions;
+package parkinglot.exceptions;
 
 public class NoSpaceInParkingLotException extends RuntimeException {
     public NoSpaceInParkingLotException(String message) {
