+++ src/test/java/parkinglot/GraduateParkingLotBoyTest.java
@@ -1,22 +1,22 @@
-package parkinglot.beta;
+package parkinglot;
 
 import org.junit.Assert;
 import org.junit.Test;
-import parkinglot.beta.exceptions.InvalidReceiptException;
-import parkinglot.beta.exceptions.NoSpaceInParkingLotException;
+import parkinglot.exceptions.InvalidReceiptException;
+import parkinglot.exceptions.NoSpaceInParkingLotException;
 
 import static java.util.Arrays.asList;
 
-public class ParkingLotBoyTest {
+public class GraduateParkingLotBoyTest {
 
     @Test
     public void should_parking_1_car_success_when_parkingboy_parking_1_car_given_2_parking_lots_with_free_space_1_2() {
         ParkingLot parkingLot = new ParkingLot(1);
         ParkingLot parkingLot2 = new ParkingLot(2);
-        ParkingLotBoy parkingLotBoy = new ParkingLotBoy(asList(parkingLot, parkingLot2));
+        GraduateParkingLotBoy graduateParkingLotBoy = new GraduateParkingLotBoy(asList(parkingLot, parkingLot2));
 
         Car car = new Car();
-        Receipt receipt = parkingLotBoy.park(car);
+        Receipt receipt = graduateParkingLotBoy.park(car);
         Assert.assertSame(car, parkingLot.pickBy(receipt));
     }
 
@@ -24,10 +24,10 @@ public class ParkingLotBoyTest {
     public void should_parking_1_car_in_the_second_success_when_parkingboy_parking_1_car_given_2_parking_lots_with_free_space_2_1() {
         ParkingLot parkingLot = new ParkingLot(2);
         ParkingLot parkingLot2 = new ParkingLot(1);
-        ParkingLotBoy parkingLotBoy = new ParkingLotBoy(asList(parkingLot, parkingLot2));
+        GraduateParkingLotBoy graduateParkingLotBoy = new GraduateParkingLotBoy(asList(parkingLot, parkingLot2));
 
         Car car = new Car();
-        Receipt receipt = parkingLotBoy.park(car);
+        Receipt receipt = graduateParkingLotBoy.park(car);
         Assert.assertSame(car, parkingLot.pickBy(receipt));
     }
 
@@ -36,10 +36,10 @@ public class ParkingLotBoyTest {
         ParkingLot parkingLot = new ParkingLot(0);
         ParkingLot parkingLot2 = new ParkingLot(1);
         ParkingLot parkingLot3 = new ParkingLot(2);
-        ParkingLotBoy parkingLotBoy = new ParkingLotBoy(asList(parkingLot, parkingLot2, parkingLot3));
+        GraduateParkingLotBoy graduateParkingLotBoy = new GraduateParkingLotBoy(asList(parkingLot, parkingLot2, parkingLot3));
 
         Car car = new Car();
-        Receipt receipt = parkingLotBoy.park(car);
+        Receipt receipt = graduateParkingLotBoy.park(car);
         Assert.assertSame(car, parkingLot2.pickBy(receipt));
     }
 
@@ -47,30 +47,30 @@ public class ParkingLotBoyTest {
     public void should_parking_1_car_failed_when_parkingboy_parking_1_car_given_2_parking_lots_with_free_space_0_0() {
         ParkingLot parkingLot = new ParkingLot(0);
         ParkingLot parkingLot2 = new ParkingLot(0);
-        ParkingLotBoy parkingLotBoy = new ParkingLotBoy(asList(parkingLot, parkingLot2));
+        GraduateParkingLotBoy graduateParkingLotBoy = new GraduateParkingLotBoy(asList(parkingLot, parkingLot2));
 
         Car car = new Car();
-        parkingLotBoy.park(car);
+        graduateParkingLotBoy.park(car);
     }
 
     @Test
     public void should_picking_1_car_success_when_parkingboy_picking_1_car_given_1_no_free_space_parking_lot() {
         ParkingLot parkingLot = new ParkingLot(1);
-        ParkingLotBoy parkingLotBoy = new ParkingLotBoy(asList(parkingLot));
+        GraduateParkingLotBoy graduateParkingLotBoy = new GraduateParkingLotBoy(asList(parkingLot));
 
         Car car = new Car();
-        Receipt receipt = parkingLotBoy.park(car);
-        Assert.assertSame(car, parkingLotBoy.pickBy(receipt));
+        Receipt receipt = graduateParkingLotBoy.park(car);
+        Assert.assertSame(car, graduateParkingLotBoy.pickBy(receipt));
     }
 
     @Test(expected = InvalidReceiptException.class)
     public void should_picking_1_car_failed_when_parkingboy_picking_1_car_twice_given_1_valid_receipt() {
         ParkingLot parkingLot = new ParkingLot(1);
-        ParkingLotBoy parkingLotBoy = new ParkingLotBoy(asList(parkingLot));
+        GraduateParkingLotBoy graduateParkingLotBoy = new GraduateParkingLotBoy(asList(parkingLot));
 
         Car car = new Car();
-        Receipt receipt = parkingLotBoy.park(car);
-        Assert.assertSame(car, parkingLotBoy.pickBy(receipt));
-        parkingLotBoy.pickBy(receipt);
+        Receipt receipt = graduateParkingLotBoy.park(car);
+        Assert.assertSame(car, graduateParkingLotBoy.pickBy(receipt));
+        graduateParkingLotBoy.pickBy(receipt);
     }
 }
