+++ src/test/java/parkinglot/SmartParkingLotBoyTest.java
@@ -1,0 +1,4 @@
+package parkinglot;
+
+public class SmartParkingLotBoyTest {
+}
